% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cmdscale_lanczos.R
\name{cmdscale_lanczos}
\alias{cmdscale_lanczos}
\title{Faster multi-dimensional scaling}
\usage{
cmdscale_lanczos(d, k = 2, eig = FALSE, add = FALSE, x.ret = FALSE)
}
\arguments{
\item{d}{a distance structure such as that returned by 'dist' or a full symmetric matrix containing the dissimilarities.}

\item{k}{the maximum dimension of the space which the data are to be represented in; must be in \code{\{1, 2, ..., n-1\}}.}

\item{eig}{indicates whether eigenvalues should be returned.}

\item{add}{logical indicating if an additive constant c* should be computed, and added to the non-diagonal dissimilarities such that the modified dissimilarities are Euclidean.}

\item{x.ret}{indicates whether the doubly centred symmetric distance matrix should be returned.}
}
\value{
as \code{\link{cmdscale}}
}
\description{
This is a stand-in replacement for \code{\link{cmdscale}} that uses the Lanczos procedure (\code{slanczos}) in \code{\link{mgcv}} instead of \code{eigen}.
}
\details{
Much of the code is based on \code{cmdscale}.
}
\author{
David L Miller
}

