% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pco_predict_preprocess.R
\name{pco_predict_preprocess}
\alias{pco_predict_preprocess}
\title{Make predictions using the pco basis}
\usage{
pco_predict_preprocess(model, newdata = NULL, dist_list)
}
\arguments{
\item{model}{a fitted model with at least one term of class \code{pco.smooth}}

\item{newdata}{prediction data}

\item{dist_list}{a list of distance matrices, each of which has distances from observations to prediction points. Each matrix should have number of rows identical to the number of observations in the model and as many columns as there are prediction points. List entries need the name of the term they refer to as their element name (e.g., if the term is \code{s(x)}, then the corresponding matrix element will be named "\code{x}").}
}
\value{
a \code{data.frame}, either the \code{newdata} augmented with new columns or a \code{data.frame} with just those columns.
}
\description{
Making predictions using the \code{pco} basis requires some pre-processing of the data first. The function \code{pco_predict_preprocess} builds a \code{data.frame} (or augments an existing one) to be used with the usual \code{predict} function.
}
\details{
The function takes uses the distances from the data in the model to the prediction points and uses Gower's interpolation to "insert" the prediction points into the existing multidimensional scaling projection. These new coordinates are returned.
}
\author{
David L Miller
}
\references{
Gower, J. C. (1968). Adding a point to vector diagrams in multivariate analysis. Biometrika, 55(3), 582. http://doi.org/10.2307/2334268

Miller, D. L. (2012). On Smooth Models for Complex Domains and Distances. (S. N. Wood, Ed.). Retrieved from http://opus.bath.ac.uk/31800/
}

