% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Predict.matrix.pco.smooth.R
\name{Predict.matrix.pco.smooth}
\alias{Predict.matrix.pco.smooth}
\title{PCO smoothing prediction}
\usage{
\method{Predict.matrix}{pco.smooth}(object, data)
}
\arguments{
\item{object}{a smooth specification object, usually generated by a term \code{s(..., bs="pco", ...)}. Note that \code{xt} object is needed, see Details.}

\item{data}{a list containing just the data (including any \code{by} variable) required by this term, with names corresponding to \code{object$term} (and \code{object$by}). The \code{by} variable is the last element.}
}
\description{
Prediction method for \code{pco} smoothers. Called via \code{\link{predict.gam}}.
}
\details{
NOTE: Predictions can only be made if a distance matrix (\code{D}) was supplied during the model setup (i.e. it doesn't work if you suppied \code{k} in \code{xt}).
}
\author{
David L Miller
}

