% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/smooth.construct.pco.smooth.spec.R
\name{smooth.construct.pco.smooth.spec}
\alias{pco}
\alias{smooth.construct.pco.smooth.spec}
\title{Build a PCO smoother}
\usage{
\method{smooth.construct}{pco.smooth.spec}(object, data, knots)
}
\arguments{
\item{object}{a smooth specification object, usually generated by a term \code{s(...,bs="pco",...)}. Note that \code{xt} object is needed, see Details, below.}

\item{data}{a list containing just the data (including any \code{by} variable) required by this term, with names corresponding to \code{object$term} (and \code{object$by}). The \code{by} variable is the last element.}

\item{knots}{IGNORED!}
}
\value{
An object of class \code{pco.smooth}.
}
\description{
Note that the usual \code{k} argument to \code{\link{s}} will set the projection dimension.
}
\section{Details}{

The constructor is not normally called directly, but is rather used internally by \code{\link{gam}}. To use for basis setup it is recommended to use \code{\link{smooth.construct2}}.

The basis dimension \code{k} will be used as the truncation for the multidimensional scaling.

When specifying the model extra arguments must be supplied by the \code{xt} argument. Two forms are possible:

First supplying:
\tabular{ll}{
  \code{realdata} \tab the actual data to use for the model, usual args are ignored\cr
  \code{dist_fn} \tab distance function, takes one arg a \code{data.frame} or \code{matrix}, returns a square distance matrix with \code{nrow(arg)} rows and cols.
}
Or instead supplying: \code{D} a distance matrix.
}
\author{
David L Miller, based on code from Lan Huo and Phil Reiss
}

